本地方法调用

+++ src/main/java/com/lgb/instructions/stack/DUP.java
@@ -6,7 +6,7 @@ import com.lgb.rtda.Frame;
 public class DUP extends NoOperandsInstruction {
     @Override
     public void execute(Frame frame) {
-        byte slot = frame.operandStack.popSlot();
+        int slot = frame.operandStack.popSlot();
         frame.operandStack.pushSlot(slot);
         frame.operandStack.pushSlot(slot);
     }
