+++ src/main/java/com/lgb/instructions/stack/DUP_X1.java
@@ -17,8 +17,8 @@ public class DUP_X1 extends NoOperandsInstruction {
     @Override
     public void execute(Frame frame) {
         OperandStack stack = frame.operandStack;
-        byte slot1 = stack.popSlot();
-        byte slot2 = stack.popSlot();
+        int slot1 = stack.popSlot();
+        int slot2 = stack.popSlot();
         stack.pushSlot(slot1);
         stack.pushSlot(slot2);
         stack.pushSlot(slot1);
