+++ src/main/java/com/lgb/instructions/stack/DUP_X2.java
@@ -18,9 +18,9 @@ public class DUP_X2 extends NoOperandsInstruction {
     @Override
     public void execute(Frame frame) {
         OperandStack stack = frame.operandStack;
-        byte slot1 = stack.popSlot();
-        byte slot2 = stack.popSlot();
-        byte slot3 = stack.popSlot();
+        int slot1 = stack.popSlot();
+        int slot2 = stack.popSlot();
+        int slot3 = stack.popSlot();
         stack.pushSlot(slot1);
         stack.pushSlot(slot3);
         stack.pushSlot(slot2);
